favorite-modify

+++ app/controllers/favorites_controller.rb
@@ -10,9 +10,10 @@ class FavoritesController < ApplicationController
 
   def destroy
     @book = Book.find(params[:book_id])
-    favorite = Favorite.find_by(book_id:@book.id)
-    favorite.user_id = current_user.id
-    favorite.destroy
+    favorite = Favorite.find_by(book_id: @book.id, user_id: current_user.id)
+    if favorite.user_id == current_user.id
+      favorite.destroy
+    end
     # redirect_to request.referer
   end
